Refactor Android app to use 'Survivor' namespace and update dependencies; implement WebView in MainActivity and adjust layout files.
