+++ app_android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">JEB</string>
+    <string name="app_name">Survivor</string>
 </resources>
