+++ app_android/settings.gradle.kts
@@ -19,5 +19,5 @@ dependencyResolutionManagement {
     }
 }
 
-rootProject.name = "JEB"
+rootProject.name = "Survivor"
 include(":app")
